# lib/api-zod/src/generated/types/answerResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerResult {
  correct: boolean;
  correctAnswer: number;
  correctAnswerText: string;
  newScore: number;
}