# lib/api-zod/src/generated/types/authCredentials.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthCredentials {
  username: string;
  password: string;
}