# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  sessionId: number;
  questionId: number;
  text: string;
  choices: string[];
  category: string;
  hasAnswered: boolean;
  answeredCount: number;
}