# lib/api-zod/src/generated/types/quizStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizStatsDistribution } from "./quizStatsDistribution";

export interface QuizStats {
  totalAnswers: number;
  correctAnswers: number;
  distribution: QuizStatsDistribution;
}